devops-geekweek/projectcode/requirements.txt
MarkupSafe==0.23
Pillow==2.7.0
PyMemoize==1.0.1
TwitterAPI==2.2.8
anyjson==0.3.3
arrow==0.5.0
backports.ssl-match-hostname==3.4.0.2
boto==2.36.0
click==3.3
config==0.3.9
dweepy==0.2.0
itsdangerous==0.24
oauthlib==0.7.2
python-dateutil==2.4.0
pytz==2014.10
redis==2.10.3
requests==2.5.1
requests-oauthlib==0.4.2
rq==0.5.0
rq-cli==0.1.2
rq-dashboard==0.3.4
simplejson==3.6.5
six==1.9.0
times==0.7
wsgiref==0.1.2
